server/requirements.txt
crewai
langchain
langchain-community
langchain-openai
faiss-cpu
openai
PyPDF2
python-decouple
fastapi
uvicorn
python-multipart
boto3 